+++ scripts/macros.js
@@ -200,6 +200,7 @@ import {investitureOfFlame} from './macros/spells/investitureOfFlame.js';
 import {investitureOfIce} from './macros/spells/investitureOfIce.js';
 import {investitureOfStone} from './macros/spells/investitureOfStone.js';
 import {investmentOfTheChainMaster} from './macros/classFeatures/warlock/invocations/investmentOfTheChainMaster.js';
+import {javelinOfLightning} from './macros/items/javelinOfLightning.js';
 import {kobold} from './macros/monsterFeatures/kobold/kobold.js';
 import {labyrinthineRecall} from './macros/raceFeatures/minotaur/labyrinthineRecall.js';
 import {lanternOfRevealing} from './macros/items/lanternOfRevealing.js';
@@ -223,6 +224,7 @@ import {metallicBreathWeapon} from './macros/raceFeatures/dragonborn/metallic/me
 import {mightyImpel} from './macros/classFeatures/barbarian/giant/mightyImpel.js';
 import {mirrorImage} from './macros/spells/mirrorImage.js';
 import {moonbeam} from './macros/spells/moonbeam.js';
+import {moteOfPotential} from './macros/classFeatures/bard/collegeOfCreation/moteOfPotential.js';
 import {mudMephit} from './macros/monsterFeatures/mudMephit/mudMephit.js';
 import {muddledMind} from './macros/classFeatures/bloodHunter/bloodCurses/muddledMind.js';
 import {mutagencraft} from './macros/classFeatures/bloodHunter/orderOfTheMutant/mutagencraft.js';
@@ -353,6 +355,10 @@ import {wrathOfTheStorm} from './macros/classFeatures/cleric/tempestDomain/wrath
 import {zealousPresence} from './macros/classFeatures/barbarian/zealot/zealousPresence.js';
 import {zombie} from './macros/monsterFeatures/zombie/zombie.js';
 import {zoneOfTruth} from './macros/spells/zoneOfTruth.js';
+import {controlUndead} from './macros/classFeatures/paladin/oathbreaker/controlUndead.js';
+import {dreadfulAspect} from './macros/classFeatures/paladin/oathbreaker/dreadfulAspect.js';
+import {auraOfHate} from './macros/classFeatures/paladin/oathbreaker/auraOfHate.js';
+import {dreadLord} from './macros/classFeatures/paladin/oathbreaker/dreadLord.js';
 export async function onHitMacro(workflow) {
     if (!workflow.targets.size) return;
     workflow.targets.forEach(async token => {
@@ -489,6 +495,20 @@ async function onMoveEffect(macroName, token, selectedAura) {
         case 'protectiveLights':
             await wildSurge.protectiveLights.aura(token, selectedAura);
             break;
+        case 'auraOfHate':
+            await auraOfHate.aura(token, selectedAura);
+            break;
+        case 'dreadLord':
+            await dreadLord.aura(token, selectedAura);
+            break;
+    }
+}
+async function onMoveSpecial(name, sourceToken, targetToken) {
+    switch(name) {
+        case 'auraOfHate':
+            return await auraOfHate.special(sourceToken, targetToken);
+        default:
+            return false;
     }
 }
 async function templateTrigger(macroName, token, trigger, reason) {
@@ -685,6 +705,7 @@ export let macros = {
     'investitureOfIce': investitureOfIce,
     'investitureOfStone': investitureOfStone,
     'investmentOfTheChainMaster': investmentOfTheChainMaster,
+    'javelinOfLightning': javelinOfLightning,
     'lanternOfRevealing': lanternOfRevealing,
     'legionOfOne': legionOfOne,
     'lifeTransference': lifeTransference,
@@ -706,6 +727,7 @@ export let macros = {
     'mirrorImage': mirrorImage,
     'monster': monster,
     'moonbeam': moonbeam,
+    'moteOfPotential': moteOfPotential,
     'muddledMind': muddledMind,
     'mutagencraft': mutagencraft,
     'necroticShroud': necroticShroud,
@@ -821,5 +843,10 @@ export let macros = {
     'wrapsOffDyamak': wrapsOffDyamak,
     'wrathOfTheStorm': wrathOfTheStorm,
     'zealousPresence': zealousPresence,
-    'zoneOfTruth': zoneOfTruth
+    'zoneOfTruth': zoneOfTruth,
+    'controlUndead': controlUndead,
+    'dreadfulAspect': dreadfulAspect,
+    'auraOfHate': auraOfHate,
+    'onMoveSpecial': onMoveSpecial,
+    'dreadLord': dreadLord
 }